+++ tests/test_visitors/test_ast/conftest.py
@@ -20,8 +20,7 @@ def _maybe_set_parent(tree: ast.AST) -> ast.AST:
     """
     for statement in ast.walk(tree):
         for child in ast.iter_child_nodes(statement):
-            if not hasattr(child, 'parent'):  # noqa: Z421
-                setattr(child, 'parent', statement)
+            setattr(child, 'parent', statement)
 
     return tree